+++ app/controllers/folio/newsletter_subscriptions_controller.rb
@@ -4,9 +4,7 @@ class Folio::NewsletterSubscriptionsController < Folio::ApplicationController
   def create
     attrs = newsletter_subscription_params.merge(visit: current_visit)
     @newsletter_subscription = Folio::NewsletterSubscription.new(attrs)
-    success = @newsletter_subscription.save
-
-    Folio::NewsletterSubscriptionMailer.notification_email(@newsletter_subscription).deliver_later if success
+    @newsletter_subscription.save
 
     render html: cell('folio/newsletter_subscription_form', @newsletter_subscription, cell_options_params)
   end