+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2019-05-15
+### Added
+- `Folio::Subscribable` concern
+- `Folio::Mailchimp::SubscribeJob`
+### Removed
+- `folio:export:newsletter` rake task
+
 ## 2019-05-02
 ### Added
 - added `public?` class method to `Page` to disable access to homepage and such via pages controller